Add secure assets plugin
Assets plugin is now passed the secure and insecure configuration
properties. The secure version of the plugin will use the former when
it constructs asset paths.

The corresponding static.path property appears to be unused in the
application but a similar approach may be required if statuc
components start to get added to common.