Refactor Omniture tracking on IdentityPage
IdentityPage now includes an Omniture mixin when created
correctky. This mixin exposes methods to add omniture tracking
parameters. Those methods return an IdentityPage instance with the
tracking params included and a slightly different type. This type can
be specified in the template to mark a view as requiring Omniture
tracking (the registration view now requires a tracked IdentityPage
for example).