Added an endpoint to change platforms.
This is a bit of a shift for me as I was hoping not to have to do this.

However I get too many  complaints that the JS version is not working properly for various reasons.

Nothing is pointing at this endpoint yet, I still need to route to it.