Killed off the old style integration tests.
Article for example has been broken since 3 Sep and nobody has bothered fixing it.

These are just getting in our way, we failed to spot that assets had stopped gzipping because it was just another broken build in a sea of red.

We have been shipping for a month just ignoring these and no kittens have died.