Inline images of the first container.
A lot of the negative feedback we have been getting has been around images not loading on the fronts.

This can be due to slow connections or Javascript errors.

This inlines (i.e. `<img/>`) the images in the *first* container on the page. The rest are lazily loaded.

This should mean we get the top images regardless of connection speed or javascript errors.